backend/requirements.txt
Flask==3.0.3
flask_cors
pinecone-client==3.1.0
python-dotenv
gunicorn
pymongo[srv]==3.12.0
flask-bcrypt
pypdf
Pillow
pdf2image
pytesseract
langchain
langchain_community
langchain_openai
langchain_pinecone
langchain_text_splitters
langchain_core
langchainhub